refactor

+++ src/main/java/alexander/ivanov/analyzer/LogRecordAnalyzer.java
@@ -10,10 +10,10 @@ import java.util.stream.Stream;
 public interface LogRecordAnalyzer extends Analyzer<Stream<String>, Stream<LogRecord>> {
     @Override
      default void analyze(Stream<LogRecord> data) {
-        analyze(data, Constants.TIME_DEFAULT.getIntValue(), Constants.ACCESSIBILITY_DEFAULT.getFloatValue());
+        analyze(data, Constants.TIME_DEFAULT.getFloatValue(), Constants.ACCESSIBILITY_DEFAULT.getFloatValue());
     }
 
-    void analyze(Stream<LogRecord> data, Integer time, Float accessibility);
+    void analyze(Stream<LogRecord> data, Float time, Float accessibility);
 
     List<ResultRecord> getResult();
 }
